agentic-case-study/usecase-003/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.11
openai>=1.3.0
python-dotenv>=1.0.0
duckduckgo-search>=3.8.3